Corrected StatesHistoryRepositoryIT and created StatesRepository
