+++ persistence/src/test/resources/sql/states/states.sql
@@ -22,22 +22,30 @@ INSERT INTO applications (candidate_id, position_id, channel_id, creation_date)
         AND p.name = 'Fejlesztő'
         AND ch.name = 'facebook';
 
-INSERT INTO states (application_id, state_index)
+INSERT INTO states_history (application_id, creation_date, state_id)
     VALUES ((SELECT id FROM applications a WHERE a.candidate_id = (SELECT id FROM candidates c WHERE c.name = 'Candidate B')
                                                  AND a.channel_id = (SELECT id FROM channels ch WHERE ch.name = 'direkt')
-                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')), 1);
+                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')),
+            NOW(),
+            (SELECT id FROM states WHERE name = 'firstTest'));
 
-INSERT INTO states (application_id, state_index)
+INSERT INTO states_history (application_id, creation_date, state_id)
     VALUES ((SELECT id FROM applications a WHERE a.candidate_id = (SELECT id FROM candidates c WHERE c.name = 'Candidate C')
                                                  AND a.channel_id = (SELECT id FROM channels ch WHERE ch.name = 'facebook')
-                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')), 1);
+                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')),
+            ADDDATE(NOW(), 2),
+            (SELECT id FROM states WHERE name = 'hr'));
 
-INSERT INTO states (application_id, state_index)
+INSERT INTO states_history (application_id, creation_date, state_id)
     VALUES ((SELECT id FROM applications a WHERE a.candidate_id = (SELECT id FROM candidates c WHERE c.name = 'Candidate C')
                                                  AND a.channel_id = (SELECT id FROM channels ch WHERE ch.name = 'facebook')
-                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')), 5);
+                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')),
+            ADDDATE(NOW(), 3),
+            (SELECT id FROM states WHERE name = 'coding'));
 
-INSERT INTO states (application_id, state_index)
+INSERT INTO states_history (application_id, creation_date, state_id)
     VALUES ((SELECT id FROM applications a WHERE a.candidate_id = (SELECT id FROM candidates c WHERE c.name = 'Candidate C')
                                                  AND a.channel_id = (SELECT id FROM channels ch WHERE ch.name = 'facebook')
-                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')), 3);
+                                                 AND a.position_id = (SELECT id FROM positions p WHERE p.name = 'Fejlesztő')),
+            ADDDATE(NOW(), 1),
+            (SELECT id FROM states WHERE name = 'firstTest'));
